requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
dateutils==0.6.12
decopatch==1.4.10
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
makefun==1.15.0
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
termcolor==2.1.1
tomli==2.0.1
typeguard==2.13.3
urllib3==1.26.13
valid8==5.1.2
